server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
boto3==1.34.0
livekit==1.0.12
pydub==0.25.1
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0 